requirements.txt
numpy==1.21.2
streamlit==1.14.0
face_recognition==1.3.0
Pillow==8.2.0